+++ CHANGELOG.rst
@@ -4,7 +4,10 @@ CHANGELOG
 5.3.63 (unreleased)
 -------------------
 
-- Update Transaction.contains, Transaction.tpc_commit
+- Check deleted list in Transaction.contains
+  [qiwn]
+
+- Change commit order (delete before add) in Transaction.tpc_commit
   [qiwn]